+++ src/test/java/org/apache/commons/io/IOUtilsWriteTestCase.java
@@ -51,7 +51,7 @@ public class IOUtilsWriteTestCase {
 
     private static final int FILE_SIZE = 1024 * 4 + 1;
 
-    private final byte[] inData = TestUtils.generateTestData((long) FILE_SIZE);
+    private final byte[] inData = TestUtils.generateTestData(FILE_SIZE);
 
     // ----------------------------------------------------------------
     // Tests
